requirements.txt
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.6
fastapi==0.63.0
h11==0.12.0
iniconfig==1.1.1
kombu==5.0.2
packaging==20.9
pika==0.13.0
pluggy==0.13.1
prompt-toolkit==3.0.16
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
redis==3.5.3
six==1.15.0
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
mysqlclient==2.0.3
